refactor(server): move metrics middleware into infra::metrics
`infra/request_id.rs`, `infra/org_host.rs`, and `infra/i18n.rs` each define
their own middleware and let `router.rs` wire it with `.layer(...)`.
`metrics_middleware` was the exception, defined inline in the router while
`infra/metrics.rs` held the recorder, the `/metrics` handler, and the
`record_*` helpers.

Moving it puts every label-producing site in one module, so "is every label
value drawn from a finite set?" is answerable by reading one file. The module
doc now states that invariant and why it matters: nothing installs an idle
timeout, so a series that gets created is never evicted.

The layer-ordering tests stay in `router.rs` -- they pin how `build_app`
composes the stack, not what the middleware records.
